+++ unlock-app/src/components/content/certification/CertificationDetails.tsx
@@ -39,7 +39,6 @@ interface CertificationDetailsProps {
   network: number
   tokenId?: string
   expiration?: string | number
-  isLoading?: boolean
 }
 
 const CertificationManagerOptions = ({
@@ -125,7 +124,6 @@ export const CertificationDetails = ({
   lockAddress,
   network,
   tokenId,
-  isLoading,
 }: CertificationDetailsProps) => {
   const { account } = useAuth()
   const config = useConfig()
@@ -170,8 +168,7 @@ export const CertificationDetails = ({
     isLockDataLoading ||
     (isMetadataLoading && !isFetched) ||
     isLoadingLockManager ||
-    isHasValidKeyLoading ||
-    isLoading
+    isHasValidKeyLoading
 
   const { isCertification } = getLockTypeByMetadata(metadata)
